refactor(app/inbound): request counting metrics submodule
